bot/requirements.txt
python-telegram-bot[job-queue]==21.*
python-dotenv>=1.0
